+++ _components/dialogmodal/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Frend Dialog Modal Change Log
 
+## [1.0.2] 2016-06-05
+- Replace NodeList.prototype jacking with `_q()` querySelectorAll util
+
 ## [1.0.1] 2016-06-04
 - Replaced aria-describedby with aria-labelledby
 - Updated initial focus to first focusable item